refactor: prisma 폴더를 api 하위 lib 폴더로 이동
