+++ app/api/kits/route.ts
@@ -1,8 +1,8 @@
 import { NextResponse } from 'next/server';
 import { auth } from '@/auth';
-import { kitSelect, prisma } from '@/lib/prisma';
+import { kitSelect, prisma } from '@/app/api/lib/prisma';
 import { S3Manager } from '@/lib/services/s3';
-import { BLURRED_IMAGE_INDEX, REWARD_IMAGE_INDEX, THUMBNAIL_IMAGE_INDEX } from '../lib/constants';
+import { BLURRED_IMAGE_INDEX, REWARD_IMAGE_INDEX, THUMBNAIL_IMAGE_INDEX } from '@/app/api/lib/constants';
 
 export async function GET(request: Request) {
   try {
