+++ app/api/me/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { prisma } from '@/lib/prisma';
+import { prisma } from '@/app/api/lib/prisma';
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
