+++ app/api/rallies/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { prisma, kitSelect, userSelect } from '@/lib/prisma';
+import { prisma, kitSelect, userSelect } from '@/app/api/lib/prisma';
 
 export async function POST(request: Request) {
   // TODO: auth, 필수 항목 검증 미들웨어 구현
