+++ auth.ts
@@ -1,7 +1,7 @@
 import NextAuth from 'next-auth';
 import Twitter from 'next-auth/providers/twitter';
 import { PrismaAdapter } from '@auth/prisma-adapter';
-import { prisma } from '@/lib/prisma';
+import { prisma } from '@/app/api/lib/prisma';
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
   adapter: PrismaAdapter(prisma),
